Move TrackerType out of imports package
