+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/imports/domain/TrackerDto.java
@@ -27,7 +27,7 @@
  */
 package org.hisp.dhis.tracker.imports.domain;
 
-import org.hisp.dhis.tracker.imports.TrackerType;
+import org.hisp.dhis.tracker.TrackerType;
 
 /**
  * @author Morten Svanæs <msvanaes@dhis2.org>
