+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/imports/validation/Warning.java
@@ -28,7 +28,7 @@
 package org.hisp.dhis.tracker.imports.validation;
 
 import lombok.Value;
-import org.hisp.dhis.tracker.imports.TrackerType;
+import org.hisp.dhis.tracker.TrackerType;
 
 @Value
 public class Warning implements Validation {
